refactor(suricata): move protocol const to package suricata/data/protocol
